pyproject.toml
[project]
name = "chofufesta-2024"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gtts>=2.5.3",
    "nfcpy>=1.0.4",
    "pygame>=2.6.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
]
